backend/requirements.txt
Django==4.2.11
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
djangorestframework==3.14.0
